split dropdowns / panels when possible
