+++ web/locales/en/plugin__netobserv-plugin.json
@@ -45,12 +45,13 @@
   "Dropped packets": "Dropped packets",
   "DNS latencies": "DNS latencies",
   "RTT": "RTT",
+  "Display options": "Display options",
   "The level of details represented.": "The level of details represented.",
   "Scope": "Scope",
   "Long labels can reduce visibility.": "Long labels can reduce visibility.",
   "Truncate labels": "Truncate labels",
   "Single graph focus": "Single graph focus",
-  "Display options": "Display options",
+  "Query options": "Query options",
   "Conversation": "Conversation",
   "Flow": "Flow",
   "Loki": "Loki",
@@ -87,7 +88,6 @@
   "Depending on the matching and filter settings, several queries can be performed under the cover, each with this parameter set, resulting in more results after aggregation.": "Depending on the matching and filter settings, several queries can be performed under the cover, each with this parameter set, resulting in more results after aggregation.",
   "Top / Bottom": "Top / Bottom",
   "Limit": "Limit",
-  "Query options": "Query options",
   "Refresh off": "Refresh off",
   "15 seconds": "15 seconds",
   "30 seconds": "30 seconds",
@@ -298,6 +298,7 @@
   "Name": "Name",
   "IP": "IP",
   "No information available for this content. Change scope to get more details.": "No information available for this content. Change scope to get more details.",
+  "Cluster name": "Cluster name",
   "Stats": "Stats",
   "Top 5 DNS latency": "Top 5 DNS latency",
   "Top 5 flow RTT": "Top 5 flow RTT",
@@ -311,7 +312,6 @@
   "Average rate": "Average rate",
   "Latest time": "Latest time",
   "Latest rate": "Latest rate",
-  "Cluster name": "Cluster name",
   "Edge": "Edge",
   "Drops": "Drops",
   "Query is slow": "Query is slow",
