+++ web/src/components/dropdowns/__tests__/metric-type-dropdown.spec.tsx
@@ -1,7 +1,7 @@
 import { mount, shallow } from 'enzyme';
 import * as React from 'react';
 
-import MetricTypeDropdown from '../metric-type-dropdown';
+import { MetricTypeDropdown } from '../metric-type-dropdown';
 
 describe('<MetricDropdown />', () => {
   const props = {
