+++ web/src/components/dropdowns/overview-display-dropdown.tsx
@@ -1,16 +1,14 @@
-import { Select, Switch, Text, TextVariants, Tooltip } from '@patternfly/react-core';
-import { InfoAltIcon } from '@patternfly/react-icons';
+import { Select, Text, TextVariants } from '@patternfly/react-core';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { FlowScope } from '../../model/flow-query';
-
 import './overview-display-dropdown.css';
-import ScopeDropdown from './scope-dropdown';
-import TruncateDropdown, { TruncateLength } from './truncate-dropdown';
+import { OverviewDisplayOptions } from './overview-display-options';
+import { TruncateLength } from './truncate-dropdown';
 
 export type Size = 's' | 'm' | 'l';
 
-export interface OverviewDisplayOptionsProps {
+export interface OverviewDisplayDropdownProps {
   metricScope: FlowScope;
   setMetricScope: (s: FlowScope) => void;
   truncateLength: TruncateLength;
@@ -20,7 +18,7 @@ export interface OverviewDisplayOptionsProps {
   allowedScopes: FlowScope[];
 }
 
-export const OverviewDisplayOptions: React.FC<OverviewDisplayOptionsProps> = ({
+export const OverviewDisplayDropdown: React.FC<OverviewDisplayDropdownProps> = ({
   metricScope,
   setMetricScope,
   truncateLength,
@@ -29,62 +27,6 @@ export const OverviewDisplayOptions: React.FC<OverviewDisplayOptionsProps> = ({
   setFocus,
   allowedScopes
 }) => {
-  const { t } = useTranslation('plugin__netobserv-plugin');
-
-  return (
-    <>
-      <div className="pf-c-select__menu-group">
-        <Tooltip content={t('The level of details represented.')}>
-          <div className="pf-c-select__menu-group-title">
-            <Text component={TextVariants.p}>
-              {t('Scope')} <InfoAltIcon />
-            </Text>
-          </div>
-        </Tooltip>
-        <div className="display-dropdown-padding">
-          <ScopeDropdown
-            data-test="scope"
-            id="scope"
-            selected={metricScope}
-            setScopeType={setMetricScope}
-            allowedScopes={allowedScopes}
-          />
-        </div>
-      </div>
-      <div className="pf-c-select__menu-group">
-        <Tooltip content={t('Long labels can reduce visibility.')}>
-          <div className="pf-c-select__menu-group-title">
-            <Text component={TextVariants.p}>
-              {t('Truncate labels')} <InfoAltIcon />
-            </Text>
-          </div>
-        </Tooltip>
-        <div className="display-dropdown-padding">
-          <TruncateDropdown id="truncate" selected={truncateLength} setTruncateLength={setTruncateLength} />
-        </div>
-      </div>
-      <div className="pf-c-select__menu-group">
-        <Switch
-          id="focus-switch"
-          className="display-dropdown-padding"
-          label={t('Single graph focus')}
-          isChecked={focus}
-          onChange={setFocus}
-        />
-      </div>
-    </>
-  );
-};
-
-export const OverviewDisplayDropdown: React.FC<{
-  metricScope: FlowScope;
-  setMetricScope: (s: FlowScope) => void;
-  truncateLength: TruncateLength;
-  setTruncateLength: (v: TruncateLength) => void;
-  focus: boolean;
-  setFocus: (v: boolean) => void;
-  allowedScopes: FlowScope[];
-}> = ({ metricScope, setMetricScope, truncateLength, setTruncateLength, focus, setFocus, allowedScopes }) => {
   const { t } = useTranslation('plugin__netobserv-plugin');
   const [isOpen, setOpen] = React.useState<boolean>(false);
